Fix send-file-descriptor, receive-file-descriptor
The msghdr struct initialized no iovecs when preparing to send or
receive file descriptors. On my machine, when the
client calls `sendmsg()` with no iobuf, it succeeds (return val is not
-1), though the receiving side receives no message, therefore failing to
send a file descriptor. I presume this is because it "optimizes" an
"empty" message away as no message, therefore still calling it a
success.

This change initializes an iovec with a size of 1 byte, the bare minimum
required to create a successfully sending message.

Tested on Linux 5.10.10 x86_64